backend/requirements.txt
# Python dependencies — all free, open source
playwright>=1.44.0
httpx>=0.27.0
pydantic>=2.0.0
ddgs>=6.1.0
spacy>=3.7.0
google-genai>=0.7.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
aiosqlite>=0.20.0
